rename trustedendpoints
