+++ kusto/conn.go
@@ -22,7 +22,7 @@ import (
 	v1 "github.com/Azure/azure-kusto-go/kusto/internal/frames/v1"
 	v2 "github.com/Azure/azure-kusto-go/kusto/internal/frames/v2"
 	"github.com/Azure/azure-kusto-go/kusto/internal/response"
-	truestedEndpoints "github.com/Azure/azure-kusto-go/kusto/trusted_endpoints"
+	truestedEndpoints "github.com/Azure/azure-kusto-go/kusto/trustedendpoints"
 	"github.com/google/uuid"
 )
