+++ kusto/test/etoe/etoe_test.go
@@ -4,11 +4,6 @@ import (
 	"context"
 	"encoding/csv"
 	"fmt"
-	"github.com/Azure/azure-kusto-go/kusto/utils"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-	"github.com/shopspring/decimal"
-	"go.uber.org/goleak"
 	"io"
 	"math/rand"
 	"net/http"
@@ -19,6 +14,12 @@ import (
 	"testing"
 	"time"
 	"unicode"
+
+	"github.com/Azure/azure-kusto-go/kusto/utils"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/shopspring/decimal"
+	"go.uber.org/goleak"
 
 	"github.com/Azure/azure-kusto-go/kusto"
 	"github.com/Azure/azure-kusto-go/kusto/data/errors"
