+++ kusto/trustedendpoints/trusted_endpoints.go
@@ -1,4 +1,4 @@
-package truestedEndpoints
+package trustedendpoints
 
 import (
 	_ "embed"
