+++ kusto/trustedendpoints/trusted_endpoints_test.go
@@ -1,4 +1,4 @@
-package truestedEndpoints
+package trustedendpoints
 
 import (
 	"fmt"
