+++ kusto/trustedendpoints/well_known_kusto_endpoints.json
